Add script for copying data.
Using Jupyter notebooks is all well and fine, but a command line tool is a little better. The command line tool accepts a list of files that will be expanded and each item put to S3. No configuration override for bucket name at this point, but not a big deal to add later.